+++ app/src/main/res/layout/activity_main.xml
@@ -8,38 +8,10 @@
     android:background="#e5e5e5"
     tools:context=".view.MainActivity">
 
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/btn_addPlayer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="20dp"
-        android:layout_marginEnd="8dp"
-        android:clickable="true"
-        android:focusable="true"
-        android:src="@android:drawable/ic_input_add"
-        android:tint="@android:color/background_light"
-        app:backgroundTint="#39b7c4"
-        app:elevation="2dp"
-        app:fabSize="normal"
-        app:layout_constraintBottom_toTopOf="@id/btn_dice"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/btn_dice"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="50dp"
-        android:layout_marginEnd="8dp"
-        android:clickable="true"
-        android:focusable="true"
-        android:src="@android:drawable/ic_menu_set_as"
-        android:tint="@android:color/background_light"
-        app:backgroundTint="@color/colorPrimary"
-        app:elevation="2dp"
-        app:fabSize="normal"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+    <FrameLayout
+        android:id="@+id/fragment-container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
 </android.support.constraint.ConstraintLayout>
